fix: metadata + jest + lychee integration for the Sentry pages
- sentry-example-page split into server page (metadata incl. noindex) +
  client test button; route added to the og-metadata EXEMPT set as an
  internal tooling page without social cards
- jest moduleNameMapper stub for @sentry/nextjs (SDK export map does not
  resolve under jest's CJS resolver when tests require() page trees)
- .lycheeignore: Sentry ingest DSN is not a browsable URL
