+++ .lycheeignore
@@ -175,3 +175,7 @@ https://support.lastpass.com/**
 # snippet — not a real URL (charities substitute their own nonprofit id).
 YOUR_NONPROFIT_ID
 
+
+# Sentry ingest DSN (client identifier in instrumentation-client.ts) -
+# not a browsable URL, envelope endpoint rejects GET
+ingest\.us\.sentry\.io
