+++ src/app/sentry-example-page/page.tsx
@@ -1,43 +1,32 @@
-'use client'
+// Sentry verification page (see the Sentry onboarding "Verify" step).
+// Server component so it can export route metadata (site standard: every
+// page carries matching openGraph/twitter fields); the error-throwing
+// button lives in the client half.
+import type { Metadata } from 'next'
+import SentryTestButton from './sentry-test-button'
 
-// Sentry verification page (see the Sentry onboarding "Verify" step):
-// clicking the button throws a deliberate error that should appear as an
-// issue in free-for-charity/javascript-nextjs within moments.
-import * as Sentry from '@sentry/nextjs'
-import { useState } from 'react'
+const title = 'Sentry Test Page'
+const description =
+  'Internal verification page for Sentry error monitoring. Clicking the button sends a deliberate test error.'
 
-class SentryExampleFrontendError extends Error {
-  constructor(message: string) {
-    super(message)
-    this.name = 'SentryExampleFrontendError'
-  }
+export const metadata: Metadata = {
+  title,
+  description,
+  // Internal tooling page - keep it out of search engines and the sitemap.
+  robots: { index: false, follow: false },
+  openGraph: { title, description },
+  twitter: { title, description },
 }
 
 export default function SentryExamplePage() {
-  const [thrown, setThrown] = useState(false)
-
   return (
     <main style={{ padding: '4rem 1rem', textAlign: 'center', fontFamily: 'sans-serif' }}>
-      <h1>Sentry Test Page</h1>
+      <h1>{title}</h1>
       <p>
         Click the button to send a test error to Sentry. If an issue appears in the free-for-charity
         project, monitoring is working.
       </p>
-      <button
-        type="button"
-        style={{ padding: '0.5rem 1.25rem', marginTop: '1rem', cursor: 'pointer' }}
-        onClick={() => {
-          setThrown(true)
-          Sentry.startSpan({ op: 'test', name: 'Sentry Example Frontend Span' }, () => {
-            throw new SentryExampleFrontendError(
-              'This error was thrown on purpose from sentry-example-page.'
-            )
-          })
-        }}
-      >
-        Throw test error
-      </button>
-      {thrown && <p style={{ marginTop: '1rem' }}>Error thrown — check Sentry Issues.</p>}
+      <SentryTestButton />
     </main>
   )
 }
